Refactor model ID handling by moving getModelId function to utils and add getMaxOutputTokens function for dynamic token limits
